% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\docType{data}
\name{anomalous_data}
\alias{anomalous_data}
\title{Theoretical scattering factors for all atomic species}
\format{A list whose elements are dataframes corresponding to atomic
        elements. Each dataframe has the following columns:
\describe{
   \item{lambda}{Specific value of wavelength in angstroms.}
   \item{f1}{Real scattering component.}
   \item{f2}{Imaginary scattering component.}
}}
\usage{
anomalous_data
}
\description{
This dataset is a list where each element is an atomic species.
Each element of the list is a dataframe with 3 columns.
}
\keyword{datasets}
