% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\docType{data}
\name{atoms}
\alias{atoms}
\title{Atom names and atomic number}
\format{The dataframe has the following columns:
\describe{
   \item{anames}{2-character string indicating the atomic name.}
   \item{Z}{Integer. The atomic number.}
}}
\usage{
atoms
}
\description{
This is a dataframe including a 2-characters string indicating the atomic
element name and an integer indicating the atomic number Z.
}
\keyword{datasets}
