% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_space.R
\name{choose_a}
\alias{choose_a}
\title{Suggests unit cell side, a, based on atom content}
\usage{
choose_a(vZ, D, k = ksigma, Ma = 5)
}
\arguments{
\item{vZ}{A vector of atom Z numbers.}

\item{D}{A real number. The distance between the two furthest atoms
in the cell.}

\item{k}{A real number. It controls the standard deviation of the 
gaussian function describing the atom and, thus, the shape of the
associated peak. The standard deviation sigma is given by:
         \code{sigma = k * sqrt(Z)}}

\item{Ma}{A real number. Each gaussian atom has tails truncated at a
distance of Ma * sigma from its peak.}
}
\value{
A real number that suggests a feasible unit cell side 
containing all atoms
}
\description{
Suggests unit cell side, a, based on atom content
}
\examples{
# 2 carbon atoms, a sulphur and an oxygen
vZ <- c(6,8,16,6)

# Distance between the two carbons is 15 angstroms
D <- 15
a <- choose_a(vZ,D)
}

