% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_space.R
\name{cold_atom_gauss}
\alias{cold_atom_gauss}
\title{Cold gaussian atom}
\usage{
cold_atom_gauss(x, a, x0 = 0, Z = 1, k = ksigma)
}
\arguments{
\item{x}{Point in the 1D cell at which this function is calculated.}

\item{a}{A real number. The width of the unit cell in which the gaussian
atom is placed.}

\item{x0}{A real number. The point corresponding to the atom's peak.}

\item{Z}{An integer number. Z is the atomic number of the atom (Z(H)=1,
Z(He)=2,Z(Li)=3,Z(B)=4, etc).}

\item{k}{A real number. It controls the standard deviation of the 
gaussian function describing the atom and, thus, the shape of the
associated peak. The standard deviation sigma is given by:
         \code{sigma = k * sqrt(Z)}}
}
\value{
A vector of length equal to the length of vector x, with
values equal to the evaluated gaussian atom.
}
\description{
Cold gaussian atom
}
\examples{
# Carbon gaussian atom in the middle of a cell
a <- 10
x0 <- 5
Z <- 6
x <- seq(0,a,length=1000)
rho <- cold_atom_gauss(x,a,x0,Z)
plot(x,rho,type="l",xlab="x",ylab=expression(rho))
}

