% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{erf}
\alias{erf}
\title{Error function for real values}
\usage{
erf(x)
}
\arguments{
\item{x}{A vector of real numbers.}
}
\value{
A real number. The integral of the gaussian, centred on zero and
   with standard deviation equal to 1, between 0 and x, multiplied by
   2/sqrt(pi).
}
\description{
Error function for real values
}
\examples{
x <- seq(-1,1,length=1000)
y <- erf(x)
plot(x,y,type="l")
}
