% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reciprocal_space.R
\name{FToRho}
\alias{FToRho}
\title{From structure factors to density}
\usage{
FToRho(a, F, hidx, N)
}
\arguments{
\item{a}{A real number. The unit cell side length.}

\item{F}{A vector of complex numbers. The structure factors corresponding to
the 1D density to be calculated.}

\item{hidx}{A vector of integer numbers. The set of 1D Miller indices
corresponding to the set of structure factors F.}

\item{N}{An integer number. The number of grid points used to calculate the
1D density.}
}
\value{
A vector of N real numbers representing the calculated 1D density at
   each of the regular N grid points.
}
\description{
Given a set of complex structure factors corresponding to a set of 1D
Miller indices, the length of the 1D unit cell, the set of Miller indices
and the number of grid points used to calculate the 1D density, this
function calculates the 1D density corresponding to the given structure
factors.
}
\examples{
# First create the crystal structure (in P1)
a <- 10
vx0 <- c(1,4,6.5)
vZ <- c(8,26,6)
vB <- c(18,20,17)
vocc <- c(1,1,1)

# Enough Fourier components (Miller indices)
hidx <- 0:20

# Compute the structure factors
F <- strufac(hidx,a,vx0,vZ,vB,vocc)

# Number of grid points
N <- 100

# Density
rtmp <- FToRho(a,F,hidx,N)

# Density plot in the unit cell
x <- rtmp$x
rho <- rtmp$rr
plot(x,rho,type="l",xlab="x",ylab=expression(rho))
}

