% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{local_maxima}
\alias{local_maxima}
\title{Find local maxima in a vector of real values}
\usage{
local_maxima(x)
}
\arguments{
\item{x}{A vector of real numbers}
}
\value{
A vector of integers corresponding to local maxima positions 
in vector x
}
\description{
Find local maxima in a vector of real values
}
\examples{
t <- seq(-10,10,length=1000)
x <- dnorm(t,mean=3)+dnorm(t,mean=7)
yM <- local_maxima(x)
}

