% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reciprocal_space.R
\name{plot_absorption_curves}
\alias{plot_absorption_curves}
\title{Plot of absorption curves}
\usage{
plot_absorption_curves(chem_el, zoom = NULL)
}
\arguments{
\item{chem_el}{1- or 2-letters character string. The chemical symbol of
interest.}

\item{zoom}{Real vector of length 2. The two values are the extremes of
the wavelength window inside which to plot the two curves. Default is
for both curves to be plotted across the full available range.}
}
\value{
Nothing, but causes a 2D plot to be displayed in a graphical window.
}
\description{
Plot f' and f'' absorption curves for the specified chemical element.
Curves can be plotted in specified wavelength regions using parameter
"zoom".
}
\examples{
# No zoom
plot_absorption_curves("Fe")

# Zoom
plot_absorption_curves("Fe",c(1,3))
}

