% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_x}
\alias{read_x}
\title{Read unit cell content (atom and coordinates).}
\usage{
read_x(filename)
}
\arguments{
\item{filename}{A character string. Existing file that includes all structural
information. The file name in general has the form "[prefix]_x.dat".}
}
\value{
A named list with the following elements:
 \itemize{
   \item{a.    Real numeric. Unit cell length in angstroms.}
   \item{SG.   SG 2-letters character string. There are only two symmetries
               possible when working within 1D crystallography, P1 (no 
               symmetry) and P-1 (inversion through the origin). SG can be 
               either "P1" or "P-1".}
   \item{data. Data frame with columns having the following names: x0 (atom
               coordinates), Z (atomic numbers), B (B-factors), 
               occ (occupancies).}
         }
}
\description{
Read unit cell length, space group, atom coordinates and all other
parameters from a coordinate file.
}
\examples{
datadir <- system.file("extdata",package="crone")
filename <- file.path(datadir,"carbon_dioxide_x.dat")
ltmp <- read_x(filename)
print(names(ltmp))
print(ltmp$data)

}

