% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_space.R
\name{reduce_to_asu}
\alias{reduce_to_asu}
\title{Reduce content of unit cell to asymmetric unit.}
\usage{
reduce_to_asu(a, vx0, vZ, vB, vocc, SG = "P1")
}
\arguments{
\item{a}{Real numeric. Unit cell length in angstroms.}

\item{vx0}{Vector of real numerics. Atom positions in the unit cell.}

\item{vZ}{Vector of integers. Atomic numbers of all atoms in the
unit cell.}

\item{vB}{Vector of real numerics. B factors for all atoms in the
unit cell.}

\item{vocc}{Vector of real numerics. Occupancies (value between 0
and 1) for all atoms in the unit cell. In this function there
is no mechanism to check whether the occupancy is appropriate in case
the atom is at a special position.}

\item{SG}{2-letters character string. There are only two symmetries
possible when working within 1D crystallography, P1 (no symmetry)
and P-1 (inversion through the origin). SG can be either "P1" or
"P-1" for this function.}
}
\value{
A named list with the following elements:
 \itemize{
   \item{vx0. Vector of real numerics indicating the reduced atomic
               positions in the asymmetric unit.}
   \item{vZ.  Vector of integers indicating the reduced atomic numbers
               for all atoms in the asymmetric unit.}
   \item{vB.  Vector of real numerics indicating the reduced B factors
               for all atoms in the asymmetric unit.}
   \item{vocc. Vector of real numerics indicating the reduced occupancies
                 for all atoms in the asymmetric unit.}
 }
}
\description{
Atom types, B factors and occupancies are restricted to those corresponding
to positions in the asymmetric unit, if the input space group (SG) is P-1.
Otherwise they are only sorted according to increasing atom positions. If
the starting positions do not correspond to a properly symmetry-expanded
set, then a warning is issued. If no symmetry is used (P1) the starting set
is forced to have values inside the unit cell and it is sorted according
to increasing atom positions.
}
\examples{
# Asymmetric unit includes 4 atoms between 0 and a/2, 1 in special position
a <- 10
vx0 <- c(1,2,4,5)
vZ <- c(6,8,6,16)
vB <- c(1,1.2,1.1,0.8)
vocc <- c(1,1,1,0.5)
ltmp <- expand_to_cell(a,vx0,vZ,vB,vocc,SG="P-1")
print(ltmp)  # Positions, atomic numbers, etc. have doubled

# Now these expanded values are used as input for reduction
ltmp2 <- reduce_to_asu(a,ltmp$vx0,ltmp$vZ,ltmp$vB,ltmp$vocc,SG="P-1")

# Compare
print(vx0)
print(ltmp2$vx0)

# SG is "P1"
a <- 16
vx0 <- c(1,2,7,9,12,16)
vZ <- c(6,6,8,8,7,6)
vB <- c(0,0,0,0,0,0)
vocc <- c(1,1,1,1,1,1)
ltmp3 <- reduce_to_asu(a,vx0,vZ,vB,vocc)
print(ltmp3) # Now all positions are inside the unit cell
}

