% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reciprocal_space.R
\name{SFobs}
\alias{SFobs}
\title{Generation of structure factors with errors}
\usage{
SFobs(h, a, vx0, vZ, vB, vocc, vx0err = NULL, ntrialP = 100,
  ntrialG = 100)
}
\arguments{
\item{h}{Real numeric. One or more 1D Miller indices.}

\item{a}{A real number. The unit cell side length.}

\item{vx0}{Vector of real numerics. Atom positions in the asymmetric
unit.}

\item{vZ}{Vector of integers. Atomic numbers of all atoms in the
asymmetric unit.}

\item{vB}{Vector of real numerics. B factors for all atoms in the
asymmetric unit.}

\item{vocc}{Vector of real numerics. Occupancies (value between 0
and 1) for all atoms in the asymmetric unit. In this function there
is no mechanism to check whether the occupancy is appropriate in case
the atom is at a special position.}

\item{vx0err}{A real number. The standard deviation of the random 
displacement of all atoms composing the structure from their correct
position. Default value is NULL, corresponding to the generation of
structure factors, with no errors, from the correct structure.}

\item{ntrialP}{Integer. The number of simulated Poisson counts for each
set of structure factor amplitudes. More counts (high ntrialP) return
smaller errors for the structure factor amplitudes.}

\item{ntrialG}{Integer. This is the number of randomly generated shifts of
each atom of the structure from its true position. The shifts follow a
gaussian distribution with mean 0 and standard deviation vx0err.}
}
\value{
A named list with two elements:
 \itemize{
   \item{F. Array of mean structure factor amplitudes, among all structure
            factor arrays simulated with specific errors.}
   \item{sF.  Array of structure factors errors. These coincide with the
              standard deviations of all structure factors arrays simulated
              with specific errors.}
        }
}
\description{
This function generates structure factors calculated starting from
the given structure and subject to two types of errors: poissonian
counting errors due to the statistical nature of the photons hitting
the crystal and normal errors due to the slight random shifting of
atoms position in all the unit cells forming the lattice.
}
\examples{

# Load thiocyanate data
datadir <- system.file("extdata",package="crone")
filename <- file.path(datadir,"thiocyanate_x.dat")
ltmp <- read_x(filename)
a <- ltmp$a
SG <- ltmp$SG
print(SG) # The structure is P1, so we don't need symmetry expansion
vx0 <- ltmp$data$x0
vZ <- ltmp$data$Z
vB <- ltmp$data$B
vocc <- ltmp$data$occ

# Miller indices used
hidx <- 1:10

# Correct amplitudes and phases
F <- strufac(hidx,a,vx0,vZ,vB,vocc)
Ftrue <- Mod(F)
phitrue <- Arg(F)

# Only poissonian errors
ltmp <- SFobs(hidx,a,vx0,vZ,vB,vocc,ntrialP=2)
print(names(ltmp))
F <- ltmp$F
Fpois <- Mod(F)

# True density
rtmptrue <- FToRho(a,complex(length.out=length(hidx),modulus=Ftrue,
 argument=phitrue),hidx,1000)
plot(rtmptrue$x,rtmptrue$rr,type="l",xlab="x",ylab=expression(rho),lwd=3)

# Density with poissonian errors
rtmppois <- FToRho(a,complex(length.out=length(hidx),modulus=Fpois,
 argument=phitrue),hidx,1000)
points(rtmppois$x,rtmppois$rr,type="l",
 lty=2,col=2,lwd=2) # Very small differences

# Only random atomic errors with standard deviation 0.3 angstroms
ltmp <- SFobs(hidx,a,vx0,vZ,vB,vocc,ntrialP=0,vx0err=0.3)
F <- ltmp$F
Fcoords <- Mod(F)

# Density with gaussian errors on atom coordinates
rtmpcoords <- FToRho(a,complex(length.out=length(hidx),
 modulus=Fcoords,argument=phitrue),hidx,1000)
points(rtmpcoords$x,rtmpcoords$rr,type="l",
 lty=3,col=3,lwd=2) # Larger differences
}

