% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_space.R
\name{structure_gauss}
\alias{structure_gauss}
\title{Structure of gaussian atoms}
\usage{
structure_gauss(x, a, vx0, vZ, vB, vocc, k = ksigma)
}
\arguments{
\item{x}{Point in the 1D cell at which this function is calculated.}

\item{a}{A real number. The width of the unit cell in which the gaussian
atom is placed.}

\item{vx0}{Vector of real numerics. Atom positions in the unit cell.}

\item{vZ}{Vector of integers. Atomic numbers of all atoms in the
unit cell.}

\item{vB}{Vector of real numerics. B factors for all atoms in the
unit cell.}

\item{vocc}{Vector of real numerics. Occupancies (value between 0
and 1) for all atoms in the unit cell.}

\item{k}{A real number. It controls the standard deviation of the 
gaussian function describing the atom and, thus, the shape of the
associated peak. The standard deviation sigma is given by:
         \code{sigma = k * sqrt(Z)}}
}
\value{
A vector of length equal to the length of vector x, with
values equal to the evaluated gaussian atom.
}
\description{
Structure formed by all gaussian atoms in the unit cell. Positions, atomic
numbers and thermal factors are given by vectors of a same length. Each
atom forming the structure is also characterised by a given occupancy (between
0 and 1).
}
\examples{
# Cell, atom types, positions and B factors
a <- 10
vx0 <- c(2,5,7)
vZ <- c(6,16,8)
vB <- c(0,0,0)

# All occupancies to 1
vocc <- c(1,1,1)

# Grid for unit cell
x <- seq(0,a,length=1000)

# Structure density
rho <- structure_gauss(x,a,vx0,vZ,vB,vocc)
plot(x,rho,type="l",xlab="x",ylab=expression(rho))

# Now reduce occupancy of sulphure
vocc[2] <- 0.5
rho <- structure_gauss(x,a,vx0,vZ,vB,vocc)
points(x,rho,type="l",col=2)

# Increase temperature of oxygen
vB[3] <- 10
rho <- structure_gauss(x,a,vx0,vZ,vB,vocc)
points(x,rho,type="l",col=3)

}

