% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reciprocal_space.R
\name{strufac}
\alias{strufac}
\title{Calculation of structure factors}
\usage{
strufac(h, a, vx0, vZ, vB, vocc, anoflag = FALSE, aK = anoK, lbda = 1,
  k = ksigma, f1f2out = TRUE)
}
\arguments{
\item{h}{Real numeric. One or more 1D Miller indices.}

\item{a}{A real number. The unit cell side length.}

\item{vx0}{Vector of real numerics. Atom positions in the asymmetric
unit.}

\item{vZ}{Vector of integers. Atomic numbers of all atoms in the
asymmetric unit.}

\item{vB}{Vector of real numerics. B factors for all atoms in the
asymmetric unit.}

\item{vocc}{Vector of real numerics. Occupancies (value between 0
and 1) for all atoms in the asymmetric unit. In this function there
is no mechanism to check whether the occupancy is appropriate in case
the atom is at a special position.}

\item{anoflag}{Logical variable. If TRUE it forces scattering factors
to include anomalous contributions. As a consequence, Friedel's
pairs will not be equal.}

\item{aK}{Real numeric. This is a fudge factor included to decrease the
strength of the anomalous contributions. Without aK the strength is too
high for 1D structures, compared to real 3D structures. So aK helps
bringing down the anomalous contribution within the 5%-10% normally
met with large-size structures. The default value is aK=0.3 (anoK is
included as internal data).}

\item{lbda}{Real numeric. This is the wavelength in angstroms. Its value
is important in relation to anomalous scattering.}

\item{k}{A real number. It controls the standard deviation of the 
gaussian function describing the atom and, thus, the shape of the
associated peak. The standard deviation sigma is given by:
      \code{sigma = k * sqrt(Z)}
The default value is k=0.05 (ksigma is included as internal data).}

\item{f1f2out}{Logical variable. This variable controls output of a small
table of f' and f'' values for all chemical elements in the structure.
Default is for the table to be printed.}
}
\value{
A vector of complex numbers, the structure factors corresponding
   to the Miller indices in input.
}
\description{
Calculates structure factors corresponding to one or more 1D Miller indices,
given the atomic content of one unit cell. Anomalous scattering can be
included using logical flag "anoflag" (default is FALSE). Crystal structures
are always considered with no symetry. Thus a 1D structure with the P-1
symmetry will have to be expanded first with function "expand_to_cell".
}
\examples{
# First create the crystal structure (P1)
a <- 10
vx0 <- c(1,4,6.5)
vZ <- c(8,26,6)
vB <- c(18,20,17)
vocc <- c(1,1,1)
lbda <- 1.7   # Close to Fe's absorption

# Miller indices (including DC (h=0) component)
hidx <- 0:10

# Now structure factors without anomalous contribution
F <- strufac(hidx,a,vx0,vZ,vB,vocc,lbda=lbda)
print(length(F))  # Includes DC component (h=0)
print(F[1])       # DC component is real
print(Mod(F))     # Amplitudes decrease with increasing
                  # resolution (Miller indices)
                  
# Now try with anomalous contributions
F <- strufac(hidx,a,vx0,vZ,vB,vocc,lbda=lbda,anoflag=TRUE)
print(F[0])  # DC component is not any longer real

}

