% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_x}
\alias{write_x}
\title{Write atomic coordinates to a file.}
\usage{
write_x(filename, vx0, vZ, vB, vocc, a, SG)
}
\arguments{
\item{filename}{A character string. Prefix of the output ASCII file to
include all structural information. The file name will be "[Prefix]_x.dat".}

\item{vx0}{Vector of real numerics. Atom positions in the asymmetric unit.}

\item{vZ}{A vector of atom Z numbers.}

\item{vB}{Vector of real numerics. B factors for all atoms in the
asymmetric unit.}

\item{vocc}{Vector of real numerics. Occupancies (value between 0
and 1) for all atoms in the asymmetric unit. In this function there
is no mechanism to check whether the occupancy is appropriate in case
the atom is at a special position.}

\item{a}{Real numeric. Unit cell length in angstroms.}

\item{SG}{2-letters character string. There are only two symmetries
possible when working within 1D crystallography, P1 (no symmetry)
and P-1 (inversion through the origin). SG can be either "P1" or
"P-1" for this function.}
}
\value{
This function does not return anything, but will create an ASCII file
   of name *_x.dat which contains all coordinates of the atoms in the structure
   and other type of information.
}
\description{
Function to export all information concerning a given structure to a
so-called coordinates file of type *_x.dat.
}
\examples{
# Create an arbitrary structure in P1
SG <- "P1"
a <- 20
vx0 <- c(2,11,16,19)
vZ <- c(6,6,16,8)
vB <- c(13,14,5,10)
vocc <- c(1,1,1,1)
prfx <- "test"
write_x(prfx,vx0,vZ,vB,vocc,a,SG)

}

